requirements.txt
customtkinter>=5.2.0
psutil>=5.9.0
pyudev>=24.0.0
groq>=0.4.0
cryptography>=41.0.0
keyring>=24.0.0
pystray>=0.19.0
Pillow>=10.0.0
requests>=2.31.0
watchdog>=3.0.0
